Added NetworkAdapter, client and server classes

+++ src/Board.java
@@ -321,7 +321,7 @@ public class Board {
 
         @Override
         public int hashCode() {
-            return Objects.hash(x, y);
+            return Objects.hash((Integer)x,(Integer) y);
         }
 
         @Override
